+++ lib/ts/recipe/thirdparty/components/themes/themeBase.tsx
@@ -20,18 +20,13 @@ import styles from "./styles.css";
 
 import type { PropsWithChildren } from "react";
 
-export const ThemeBase: React.FC<
-    PropsWithChildren<{ loadDefaultFont: boolean; userStyles: Array<string | undefined> }>
-> = ({ children, userStyles, loadDefaultFont }) => {
+export const ThemeBase: React.FC<PropsWithChildren<{ userStyles: Array<string | undefined> }>> = ({
+    children,
+    userStyles,
+}) => {
     return (
         <Fragment>
             {children}
-            {loadDefaultFont && (
-                <link
-                    href="//fonts.googleapis.com/css2?family=Rubik:wght@300;400;500;600;700"
-                    rel="stylesheet"
-                    type="text/css"></link>
-            )}
             <style>
                 {styles}
                 {userStyles.join("\n")}